[IMP] point_of_sale: Simplify category selector
Since this components is only used in PoS. Use pos service instead of
props.
