+++ addons/point_of_sale/static/src/app/components/category_selector/category_selector.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <templates id="template" xml:space="preserve">
     <t t-name="point_of_sale.CategorySelector">
-        <div t-att-class="props.class" class="d-grid product-list gap-2" t-att-style="props.style">
-            <t t-foreach="props.categories" t-as="category" t-key="category.id">
-                <button t-on-click="() => props.onClick(category.id)"
+        <div class="d-grid product-list gap-2 p-2">
+            <t t-foreach="this.getCategoriesAndSub()" t-as="category" t-key="category.id">
+                <button t-on-click="() => this.pos.setSelectedCategory(category.id)"
                     t-attf-class="o_colorlist_item_color_{{!category.isSelected and !category.isChildren ? 'transparent_': ''}}{{category.color or 'none'}}"
                     t-att-class="{'opacity-50': !category.isChildren and !category.isSelected, 'justify-content-center': ui.isSmall}"
                     class="category-button px-1 btn btn-light d-flex align-items-center rounded-3"
